LayoutTests/platform/ios/imported/w3c/web-platform-tests/css/css-pseudo/backdrop-pseudo-element-click.tentative-expected.txt


FAIL Clicking on dialog's ::backdrop sets pseudoTarget to the CSSPseudoElement assert_equals: Click inside dialog: target is the dialog expected Element node <dialog id="target" open=""></dialog> but got null
